refactor: move all builtin table logic into sqlbuiltins
